requirements.txt
Django==1.8.3
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.9.0
PyYAML==3.11
-e git://github.com/mitsuhiko/werkzeug.git@d637320cd174eec5cc1baa3aa3412bd5390e9442#egg=Werkzeug-master
argparse==1.2.1
binaryornot==0.4.0
chardet==2.3.0
click==3.3
cookiecutter==1.0.0
django-admin-bootstrapped==2.5.5
django-changuito==0.7
django-crispy-forms==1.5.1
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-taggit==0.17.0
funcsigs==0.4
mock==1.3.0
pbr==1.7.0
pydot==1.0.2
pygraphviz==1.2
pyparsing==1.5.7
requests==2.7.0
six==1.9.0
sqlparse==0.1.16
wsgiref==0.1.2
